Add rpmbuild support
This commit adds the spec file for rpm2python that is no being used
on Rutgers Koji. It also modifies the Makefile.am file to make it more
compatible with the build system (you cannot run pip in a mock chroot)
and we changed the version number to 0.8.0 (from the previous 0.8).
Note that the Makefile.am now requires a venv.tar.gz file instead of
using pip to setup the venv on install.